requirements.txt
# zipfile
boto